* Integrate mpi_engine fixture with engine_runner * Switch tests to wait on future instead of polling the queue
